+++ examples/provider-install-verification/provider.tf
@@ -1,13 +1,9 @@
 # terraform {
 #   required_providers {
 #     tilgangsportalen = {
-#       source = "test.sits.no/terraform/tilgangsportalen"
+#       source = "skatteetaten/tilgangsportalen"
 #     }
 #   }
 # }
 
-# provider "tilgangsportalen" {
-#   hosturl  = var.TILGANGSPORTALEN_URL
-#   username = var.TILGANGSPORTALEN_USERNAME
-#   password = var.TILGANGSPORTALEN_PASSWORD
-# }
+# provider "tilgangsportalen" {}