+++ internal/tilgangsportalapi/get_system_role.go
@@ -8,9 +8,9 @@ import (
 
 // GetSystemRole gets information about a specific named role / check if role
 // exists. Gets the name, description, approval level, system role owner,
-// system role security owner, product category and IsForITShop for a specific
+// system role security owner and product category for a specific
 // (named) system role.
-// See https://wiki.sits.no/display/IDABAS/19.+Get+Role
+// See https://wiki.sits.no/display/IDABAS/17.+Get+Role
 func (client *Client) GetSystemRole(roleName string) (*SystemRole, error) {
 	var data SystemRole
 	log.Printf("Fetching system role %s", roleName)